DP-33139: Changes for aws provider v5 compatability (#191)
ASG:
- Added the `AmazonECSManaged` tag to the autoscaling group. This is set
on all of our ASGs, but as of aws provider version 5, terraform wants to
remove it since it is not defined in the code anywhere.
- Added a `amazon_ecs_managed_tag` variable so the above tag can be
disabled if necessary.

ECS:
- Added a `amazon_ecs_managed_tag` variable so the `AmazonECSManaged`
tag can be disabled in the auto scaling group.

RDS:
- `rds_snapshot_delete` policy was being defined even when it wasn't
used. This made the plan somewhat confusing when I was dealing with
other upgrade issues. I've fixed it so that it only defines the policy
when it will be used.
- `aws_db_instance.id` changes im aws provider v5. It no longer
references the `identifier` attribute. I changed all uses of the `id`
attribute within the module to use `identifier` instead.
- Added `rds_instance_identifier` output, which refers to the
`identifier` attribute. I did not change the `rds_instance_id` output.